move read_problem to file utils
