+++ src/problem/cvrp.py
@@ -1,8 +1,6 @@
 import math
 from dataclasses import dataclass
 
-from utils.enums import DataFileConstants as DFC
-
 
 class Cvrp:
     def __init__(self, no_of_cities, truck_capacity, cities, depot_number):
@@ -28,41 +26,6 @@ class Cvrp:
     def __str__(self):
         print(
             f"CVRP(no_of_cities={self.no_of_cities}, truck_capacity={self.truck_capacity}, depot_number={self.depot_number})")
-
-
-def read_problem(file_path: str) -> Cvrp:
-    no_of_cities = 0
-    truck_capacity = 0
-    city_coordinates = []
-    city_demand = []
-    depot_number = 0
-    with open(file_path, "r") as file:
-        for line in file:
-            if DFC.DIMENSION.value in line:
-                no_of_cities = int(line.split(':')[1].strip())
-            elif DFC.CAPACITY.value in line:
-                truck_capacity = int(line.split(':')[1].strip())
-            elif DFC.NODE_COORD_SECTION.value in line:
-                for i in range(0, no_of_cities):
-                    line = file.readline()
-                    city_coordinates.append(line.strip().split(" "))
-            elif DFC.DEMAND_SECTION.value in line:
-                for i in range(0, no_of_cities):
-                    line = file.readline()
-                    city_demand.append(line.strip().split(" "))
-            elif DFC.DEPOT_SECTION.value in line:
-                depot_number = int(file.readline().strip())
-
-    cities = []
-    for i in range(0, no_of_cities):
-        number = int(city_coordinates[i][0].strip())
-        x = int(city_coordinates[i][1].strip())
-        y = int(city_coordinates[i][2].strip())
-        demand = int(city_demand[i][1].strip())
-        city = City(number - 1, x, y, demand)  # - 1 to adjust that arrays start from 0
-        cities.append(city)
-
-    return Cvrp(no_of_cities, truck_capacity, cities, depot_number - 1)  # - 1 to adjust that arrays start from 0
 
 
 # this function needs to be refactored - take into account when truck is ie. 80% empty - maybe then turn back instead of going to depot only when its empty
