+++ src/utils/file_utils.py
@@ -3,7 +3,9 @@ import os
 from pathlib import Path
 
 from algorithm.result import Result
+from problem.cvrp import Cvrp, City
 from utils.configuration import Config
+from utils.enums import DataFileConstants as DFC
 
 root_dir = Path(__file__).resolve().parents[2]
 
@@ -34,3 +36,38 @@ def get_or_create_results_dir(problem_instance, algorithm):
     output_dir = root_dir / "results" / f"{problem_instance}/{algorithm}/{current_time}"
     os.makedirs(output_dir, exist_ok=True)
     return output_dir
+
+
+def read_problem(file_path: str) -> Cvrp:
+    no_of_cities = 0
+    truck_capacity = 0
+    city_coordinates = []
+    city_demand = []
+    depot_number = 0
+    with open(file_path, "r") as file:
+        for line in file:
+            if DFC.DIMENSION.value in line:
+                no_of_cities = int(line.split(':')[1].strip())
+            elif DFC.CAPACITY.value in line:
+                truck_capacity = int(line.split(':')[1].strip())
+            elif DFC.NODE_COORD_SECTION.value in line:
+                for i in range(0, no_of_cities):
+                    line = file.readline()
+                    city_coordinates.append(line.strip().split(" "))
+            elif DFC.DEMAND_SECTION.value in line:
+                for i in range(0, no_of_cities):
+                    line = file.readline()
+                    city_demand.append(line.strip().split(" "))
+            elif DFC.DEPOT_SECTION.value in line:
+                depot_number = int(file.readline().strip())
+
+    cities = []
+    for i in range(0, no_of_cities):
+        number = int(city_coordinates[i][0].strip())
+        x = int(city_coordinates[i][1].strip())
+        y = int(city_coordinates[i][2].strip())
+        demand = int(city_demand[i][1].strip())
+        city = City(number - 1, x, y, demand)  # - 1 to adjust that arrays start from 0
+        cities.append(city)
+
+    return Cvrp(no_of_cities, truck_capacity, cities, depot_number - 1)  # - 1 to adjust that arrays start from 0
